Change params to **kwargs, move init type checking
With this commit, `Datac` objects now require named arguments as the
constant parameters of the calculation instead of collecting these
parameters into a single dictionary. Additionally, checks on the
arguments to `__init__` were moved from the `__init__` method itself to
the corresponding class attributes. Some of the argument checking was
removed because it was unnecessary.

Closes #27.